Created an MVP for the scene.

+++ src/include/utility.h
@@ -16,11 +16,11 @@
 class Vector2 {
 
 public:
-    double x;
-    double y;
+    float x;
+    float y;
 
     Vector2() {};
-    Vector2(double x, double y) : x(x), y(y) {};
+    Vector2(float x, float y) : x(x), y(y) {};
     Vector2 operator + (const Vector2&);
     Vector2& operator = (const Vector2&);
     ~Vector2() {};
